Refactor: Move SelectedComponent into Components
