+++ indexinglanguage/src/main/java/com/yahoo/vespa/indexinglanguage/expressions/EmbedExpression.java
@@ -29,7 +29,7 @@ import java.util.Optional;
 public class EmbedExpression extends Expression  {
 
     private final Linguistics linguistics;
-    private final SelectedComponent<Embedder> embedder;
+    private final Components.Selected<Embedder> embedder;
     private final String requestedEmbedderId;
 
     /** The destination the embedding will be written to on the form [schema name].[field name] */
@@ -38,7 +38,7 @@ public class EmbedExpression extends Expression  {
     public EmbedExpression(Linguistics linguistics, Components<Embedder> embedders, String embedderId, List<String> embedderArguments) {
         this.linguistics = linguistics;
         this.requestedEmbedderId = embedderId;
-        embedder = new SelectedComponent<>("embedder", embedders, embedderId, true, embedderArguments);
+        embedder = new Components.Selected<>("embedder", embedders, embedderId, true, embedderArguments);
     }
 
     /** @return the requested embedder id. This will diverge from the selected embedder's id when executed in config-model */
