+++ indexinglanguage/src/main/java/com/yahoo/vespa/indexinglanguage/expressions/GenerateExpression.java
@@ -24,7 +24,7 @@ import java.util.Objects;
 public class GenerateExpression extends Expression {
 
     private final Linguistics linguistics;
-    private final SelectedComponent<FieldGenerator> generator;
+    private final Components.Selected<FieldGenerator> generator;
 
     /** The destination the generated value will be written to in the form [schema name].[field name] */
     private String destination;
@@ -37,8 +37,8 @@ public class GenerateExpression extends Expression {
                               String generatorId,
                               List<String> generatorArguments) {
         this.linguistics = linguistics;
-        this.generator = new SelectedComponent<>("generator", generators, generatorId, true,
-                                                 generatorArguments);
+        this.generator = new Components.Selected<>("generator", generators, generatorId, true,
+                                                   generatorArguments);
     }
 
     @Override
